feat: ask to user if is sure to delete a superhero

+++ src/app/components/heroes-list/heroes-list.component.ts
@@ -51,7 +51,9 @@ export class HeroesListComponent implements OnInit {
   }
 
   onRemove(hero: Hero) {
-    //aply verification
-    this.service.removeHero(hero.id);
+    const confirm = window.confirm('Are you sure you want to delete ' + hero.name + '?');
+    if (confirm) {
+      this.service.removeHero(hero.id);
+    }
   }
 }
